requirements.txt
tqdm
requests
aiohttp
python-dotenv
openai